\name{chromatogramMatrix}
\alias{chromatogramMatrix}
\docType{data}
\title{
chromatogram builder for m/z = 263.1678 in 003.d from cord blood sample
}
\description{
This data illustrates a chromatogram and baseline vectors to indicate chromatogram development.
}
\usage{data("chromatogramMatrix")}
\format{
  A data frame with 219 observations on the following 6 variables.
  \describe{
    \item{\code{scanNumber}}{a numeric vector}
    \item{\code{retentionTime}}{a numeric vector}
    \item{\code{smoothChromatogram}}{a numeric vector}
    \item{\code{rawChromatogram}}{a numeric vector}
    \item{\samp{12C/13C Isotopologue Pairs}}{a numeric vector}
    \item{\code{Baseline}}{a numeric vector}
  }
}
\examples{
data(chromatogramMatrix)
}
\keyword{datasets}
