\name{IPA_CompoundsAnnotation}
\alias{IPA_CompoundsAnnotation}
\title{
Compound-centric peak annotation
}
\description{
This function performs compound-centric peak annotation.
}
\usage{
IPA_CompoundsAnnotation(PARAM)
}
\arguments{
  \item{PARAM}{
a data frame from IPA_xlsxAnalyzer function containing the IPA parameters.
}
}
\value{
This function saves individual .csv files for each compound in the "compound_centric_annotation" folder.
}
