\name{IPA_logRecorder}
\alias{IPA_logRecorder}
\title{
IPA_logRecorder
}
\description{
IPA_logRecorder
}
\usage{
IPA_logRecorder(messageQuote, allowedPrinting = TRUE)
}
\arguments{
  \item{messageQuote}{
messageQuote
}
  \item{allowedPrinting}{
allowedPrinting
}
}
\value{
a line of communication messages is exported to the console and the log .txt file.
}