\name{IPA_message}
\alias{IPA_message}
\title{
IPA message
}
\description{
IPA_message
}
\usage{
IPA_message(messageQuote, failedMessage= TRUE)
}
\arguments{
  \item{messageQuote}{
messageQuote
}
  \item{failedMessage}{
failedMessage
}
}
\value{
a line of communication messages is exported to the console.
}