\name{IPA_peak_alignment_folder_xlsxAnalyzer}
\alias{IPA_peak_alignment_folder_xlsxAnalyzer}
\title{
IPA peak alignment folder xlsxAnalyzer
}
\description{
IPA peak alignment folder xlsxAnalyzer
}
\usage{
IPA_peak_alignment_folder_xlsxAnalyzer(PARAM, PARAM_ID, checkpoint_parameter,
correctedRTcheck = FALSE, CSAcheck = FALSE, allowedVerbose = TRUE)
}
\arguments{
  \item{PARAM}{
PARAM
}
  \item{PARAM_ID}{
PARAM_ID
}
  \item{checkpoint_parameter}{
checkpoint_parameter
}
  \item{correctedRTcheck}{
correctedRTcheck
}
  \item{CSAcheck}{
CSAcheck
}
  \item{allowedVerbose}{
c(TRUE, FALSE). A `TRUE` allowedVerbose provides messages about the flow of the function.
}
}
\value{
  \item{PARAM}{
PARAM
}
  \item{checkpoint_parameter}{
checkpoint_parameter
}
}
