\name{peak_spline}
\alias{peak_spline}
\docType{data}
\title{
peak spline
}
\description{
illusterates a smoothe peak using cubic spline smoothing method
}
\usage{data("peak_spline")}
\format{
  A data frame with 100 observations on the following 3 variables.
  \describe{
    \item{\code{rt_spline}}{a numeric vector}
    \item{\code{int_spline}}{a numeric vector}
    \item{\code{bl_approx}}{a numeric vector}
  }
}
\examples{
data(peak_spline)
}
\keyword{datasets}
