\name{peakPropertyTableFreqCalculator}
\alias{peakPropertyTableFreqCalculator}
\title{
Peak Property Table Frequency Calculator
}
\description{
Peak Property Table Frequency Calculator
}
\usage{
peakPropertyTableFreqCalculator(peakPropertyTable, startColumnIndex = 3,
number_processing_threads = 1, allowedVerbose = TRUE)
}
\arguments{
  \item{peakPropertyTable}{
peakPropertyTable
}
  \item{startColumnIndex}{
startColumnIndex
}
  \item{number_processing_threads}{
number_processing_threads
}
  \item{allowedVerbose}{
c(TRUE, FALSE). A `TRUE` allowedVerbose provides messages about the flow of the function.
}
}
\value{
a vector of frequency of detection.
}
