\name{peakPropertyTableMedianCalculator}
\alias{peakPropertyTableMedianCalculator}
\title{
Peak Property Table Median Calculator
}
\description{
Peak Property Table Median Calculator
}
\usage{
peakPropertyTableMedianCalculator(peakPropertyTable, falggingVector = NULL,
number_processing_threads = 1, allowedVerbose = TRUE)
}
\arguments{
  \item{peakPropertyTable}{
peakPropertyTable
}
  \item{falggingVector}{
falggingVector
}
  \item{number_processing_threads}{
number_processing_threads
}
  \item{allowedVerbose}{
c(TRUE, FALSE). A `TRUE` allowedVerbose provides messages about the flow of the function.
}
}
\value{
updated peak property table
}
