\name{peakXcolFlagger}
\alias{peakXcolFlagger}
\title{
PeakXcol Flagger
}
\description{
PeakXcol Flagger
}
\usage{
peakXcolFlagger(mzPeakXcol, rtPeakXcol, freqPeakXcol, massAccuracy, RTtolerance,
maxRedundantPeakFlagging)
}
\arguments{
  \item{mzPeakXcol}{
mzPeakXcol
}
  \item{rtPeakXcol}{
rtPeakXcol
}
  \item{freqPeakXcol}{
freqPeakXcol
}
  \item{massAccuracy}{
massAccuracy
}
  \item{RTtolerance}{
RTtolerance
}
  \item{maxRedundantPeakFlagging}{
maxRedundantPeakFlagging
}
}
\value{
a vector with flagged numbers
}