\name{plot_simple_tic}
\alias{plot_simple_tic}
\title{
plot_simple_tic
}
\description{
plot_simple_tic
}
\usage{
plot_simple_tic(filelist, filelocation, number_processing_threads = 1,
plotTitle = "Total Ion Chromatogram")
}
\arguments{
  \item{filelist}{
filelist
}
  \item{filelocation}{
filelocation
}
  \item{number_processing_threads}{
number of processing threads
}
  \item{plotTitle}{
plotTitle
}
}
\value{
plot_simple_tic
}
