\name{segment}
\alias{segment}
\docType{data}
\title{
segment
}
\description{
illusterates an output matrix of chromatogram peak dertection module from the
"chromatogram_builder.rda" object.
}
\usage{data("segment")}
\format{
  The format is:
 num [1:16, 1:2] 7 15 23 33 38 46 67 86 102 118 ...
}
\examples{
data(segment)
}
\keyword{datasets}
