\name{xlsxAnalyzer_EIC}
\alias{xlsxAnalyzer_EIC}
\title{
xlsxAnalyzer EIC
}
\description{
This function processes the spreadsheet of the IPA parameters to ensure the parameter inputs are in agreement with the IPA_EIC requirements.
}
\usage{
xlsxAnalyzer_EIC(spreadsheet)
}
\arguments{
  \item{spreadsheet}{
contains the IPA parameters.
}
}
\value{
This function returns the IPA parameters to feed the IPA_TargetedAnalysis function.
}
