\name{IPA_targeted_xlsxAnalyzer}
\alias{IPA_targeted_xlsxAnalyzer}
\title{
xlsxAnalyzer EIC
}
\description{
This function processes the spreadsheet of the IPA parameters to ensure the parameter inputs are in agreement with the `IPA_TargetedAnalysis` function.
}
\usage{
IPA_targeted_xlsxAnalyzer(spreadsheet)
}
\arguments{
  \item{spreadsheet}{
contains the IPA parameters.
}
}
\value{
This function returns the IPA parameters to feed the IPA_TargetedAnalysis function.
}
