\name{plot_simple_tic}
\alias{plot_simple_tic}
\title{
plot_simple_tic
}
\description{
plot_simple_tic
}
\usage{
plot_simple_tic(filelist, filelocation, numberOfcores,
plotTitle = "Total Ion Chromatogram")
}
\arguments{
  \item{filelist}{
filelist
}
  \item{filelocation}{
filelocation
}
  \item{numberOfcores}{
numberOfcores
}
  \item{plotTitle}{
plotTitle
}
}
\value{
plot_simple_tic
}
